app/pyproject.toml
[project]
name = "pgtester"
version = "1.0.0"
description = "Testing dummy accesses to PostgreSQL."
dependencies = [
    "flask",
    "gunicorn",
    "psycopg2-binary"
]

[build-system]
requires = ["flit_core<4"]
build-backend = "flit_core.buildapi"